Enemies and trees now spawn and have a script.
Added a game object for spawning enemies and trees.
Written a spawning script for both trees and enemies
Ensured that the speed of the enemies increases as time goes by.
Ensured that the bullets cannot affect the trees.
Added a box collider to the trees so that the player dies if he collides with it.
Written a script that makes trees move.
Making sure that enemies and trees going off the road are removed.